Styling

+++ db/migrate/20160425144901_create_instructions.rb
@@ -5,7 +5,7 @@ class CreateInstructions < ActiveRecord::Migration
       t.belongs_to :ingredient
       t.belongs_to :user, null: false
       t.string :unit, null: false
-      t.integer :quantity, null: false
+      t.float :quantity, null: false
       t.integer :add_point, null: false
       t.string :temperature
       t.string :details
